Migrated Element Handler to a Plugin Architecture
The plugin architecture should allow end-users to create alternate element types and expand the functionality of LabelWeb dynamically. Each of the built-in types are implemented as plugins and then there is a folder for placing custom plugins.

Also moved the AdjustFondToFit function to the element base so that it's available to any custom element handlers as well.