Moved form element definition to plugins
This adds new element fields to all elements, specifically form_type, required, placeholder, and description. These fields are only used in the template print page.
Does not yet support JSON elements in the preview.
Added custom element definitions for the datamatrix and imageURL components. DM is always required because if there is no data, it cannot be rendered. ImageURL should be of type URL.